5.2 Write a test for the Date class 2-3

+++ src/Streamer.h
@@ -145,9 +145,9 @@ inline my_ifstream& operator>>(my_ifstream& is, Date& v)
 {
     string tmp;
     is >> tmp;
-    unsigned y = std::atoi(tmp.substr(0, 4).c_str());
-    unsigned m = std::atoi(tmp.substr(4, 2).c_str());
-    unsigned d = std::atoi(tmp.substr(6, 2).c_str());
+    unsigned y = (unsigned int)std::atoi(tmp.substr(0, 4).c_str());
+    unsigned m = (unsigned int)std::atoi(tmp.substr(4, 2).c_str());
+    unsigned d = (unsigned int)std::atoi(tmp.substr(6, 2).c_str());
     v.init(y, m, d);
     return is;
 }
